Added the text to speech function and dart linter.
The whole app can work on Android now. It can scan and connect to the "Sight++" WiFi and initiate connection with the server, then the user can press and hold the button to speak the request out, then the app sends the request to the server and the server sends respond back to the app and the app can speak it out.